DashBoard features

+++ src/main/java/upacademy/grouproject/view/TriageBean.java
@@ -10,6 +10,7 @@ import javax.inject.Inject;
 import javax.inject.Named;
 
 import upacademy.grouproject.model.Triage;
+import upacademy.grouproject.model.TriageHistoric;
 import upacademy.grouproject.service.TriageService;
 
 @Named("triageBean")
@@ -25,6 +26,9 @@ public class TriageBean implements Serializable {
 
 	@Inject
 	private TriageService triageService;
+
+	@Inject
+	private TriageHistoricBean triageHistoricBean;
 
 	// Create triage
 	public void triageCreate(Triage triage, String nHS) {
@@ -43,32 +47,51 @@ public class TriageBean implements Serializable {
 		return triageService.sortTriage();
 	}
 
-	private List<Triage> ticketShow= new ArrayList<Triage>();
+	private List<Triage> ticketShow = new ArrayList<Triage>();
 
-	// Ticket Call
+	private List<Integer> counters = new ArrayList<Integer>();
+
+	// Dashboard Ticket Call
 	public void callTicket() {
-		triageService.deleteEntity(ticketShow.get(0).getID(), Triage.class);
-		ticketShow.clear();
-		showTickets();
+		if (triageService.consultEntity("Triage").size() == 0) {
+			// Resolver
+		} else {
+			triageService.deleteEntity(ticketShow.get(0).getID(), Triage.class);
+			triageHistoricBean.triageHistoricCreate(ticketShow.get(0));
+			ticketShow.clear();
+			counters.clear();
+			showTickets();
+			triageHistoricBean.callTicketPrev();
+		}
+
 	}
-	
+
+	// Dashboard ticket panel
 	@PostConstruct
-	public void showTickets(){
-		List<Triage> list = triageService.sortTriage();
-		Triage now = list.get(0);
-		Triage next = list.get(1);
-		
-		ticketShow.add(now);
-		ticketShow.add(next);
-		
+	public void showTickets() {
+		if (triageService.consultEntity("Triage").size() == 0) {
+			// Resolver
+		} else {
+			List<Triage> list = triageService.sortTriage();
+			Triage now = list.get(0);
+			Triage next = list.get(1);
+			ticketShow.add(now);
+			ticketShow.add(next);
+
+			counters.add(triageService.countTickets().get(0));
+			counters.add(triageService.countTickets().get(1));
+			counters.add(triageService.countTickets().get(2));
+			counters.add(triageService.countTickets().get(3));
+			counters.add(triageService.countTickets().get(4));
+		}
+
 	}
-	
-	// Count tickets 
-	public void ticketCounter () {
-		
+
+	// Count tickets
+	public void ticketCounter() {
+
 	}
-	
-	
+
 	// New triage bean
 	public void newBean() {
 		this.triage = new Triage();
@@ -118,5 +141,21 @@ public class TriageBean implements Serializable {
 
 	public void setTicketShow(List<Triage> ticketShow) {
 		this.ticketShow = ticketShow;
+	}
+
+	public TriageHistoricBean getTriageHistoricBean() {
+		return triageHistoricBean;
+	}
+
+	public void setTriageHistoricBean(TriageHistoricBean triageHistoricBean) {
+		this.triageHistoricBean = triageHistoricBean;
+	}
+
+	public List<Integer> getCounters() {
+		return counters;
+	}
+
+	public void setCounters(List<Integer> counters) {
+		this.counters = counters;
 	}
 }
